+++ src/Scene/Camera.java
@@ -4,7 +4,7 @@ import rMath.Vector3D;
 
 public class Camera extends Scene.objects.Shape {
     Vector3D direction = new Vector3D(0, 0, 1);
-    public Camera (double x, double y, double z) {
+    public Camera (float x, float y, float z) {
         super(x, y, z, 1);
 
     }
