+++ src/Scene/lighting/DirectLight.java
@@ -8,12 +8,12 @@ public class DirectLight extends Light {
     public Color ambient;
     public Vector3D direction;
 
-    public DirectLight(double x, double y, double z, Vector3D direction, Color ambient) {
+    public DirectLight(float x, float y, float z, Vector3D direction, Color ambient) {
         super(x, y, z, Color.white);
         this.direction = direction;
         this.ambient = ambient;
     }
-    public DirectLight(double x, double y, double z, Vector3D direction, Color ambient, Color color) {
+    public DirectLight(float x, float y, float z, Vector3D direction, Color ambient, Color color) {
         super(x, y, z, Color.white);
         this.direction = direction;
         this.color = color;
