+++ src/Scene/lighting/RectLight.java
@@ -3,7 +3,7 @@ package Scene.lighting;
 import java.awt.*;
 
 public class RectLight extends Light{
-    public RectLight(double x, double y, double z, Color color) {
+    public RectLight(float x, float y, float z, Color color) {
         super(x, y, z, color);
     }
 }
