+++ src/rMath/Vertex2D.java
@@ -1,22 +1,22 @@
 package rMath;
 
 public class Vertex2D {
-    public double x, y;
-    public double z; // used for ordering vertices
+    public float x, y;
+    public float z; // used for ordering vertices
 
     public Vertex2D[] edges = {};
-    public Vertex2D(double x, double y) {
+    public Vertex2D(float x, float y) {
         this.x = x;
         this.y = y;
     }
     public Vertex2D() {} // temporarily unset x an y values
 
-    public void scaleAdd(double scale, Vector3D vector, Vertex2D start_vertex) {
+    public void scaleAdd(float scale, Vector3D vector, Vertex2D start_vertex) {
         this.x = start_vertex.x + scale * vector.i;
         this.y = start_vertex.y + scale * vector.j;
     }
 
-    public void scaleAdd(double scale, Vector3D vector) {
+    public void scaleAdd(float scale, Vector3D vector) {
         scaleAdd(scale, vector, this);
     }
